test: keep responsive preview test
